backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Lender Matching Platform API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0",
    "pymupdf>=1.24.0",
    "hatchet-sdk>=0.40.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]
